requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
numpy>=1.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0

# Optional Dependencies (for future use)
pydantic-settings>=2.1.0  # For configuration management (not currently used)
httpx>=0.25.0             # For HTTP client calls (not currently used)
python-dotenv>=1.0.0      # For environment variable management (not currently used)

# Phase 2 Dependencies
watchdog>=3.0.0           # For file system monitoring and hot-reloading
PyYAML>=6.0.1             # For YAML Domain Pack support

# Phase 6 Dependencies - Database & Migrations
sqlalchemy>=2.0.0         # SQLAlchemy ORM for database models
alembic>=1.13.0          # Database migration tool
asyncpg>=0.29.0          # Async PostgreSQL driver
psycopg2-binary>=2.9.9   # PostgreSQL adapter (for Alembic sync operations)
aiosqlite>=0.19.0        # Async SQLite driver (for testing)

requests

prometheus-client
# testing
faker

# Phase 9 Dependencies - Message Broker
confluent-kafka>=2.3.0  # Preferred Kafka client library
kafka-python>=2.0.2     # Fallback Kafka client library

# Authentication
PyJWT>=2.8.0  # JWT token encoding/decoding

# FastAPI form data support
python-multipart>=0.0.6  # Required for FastAPI form data handling